+++ android/MyApplication/app/src/main/java/com/example/chun/whefe/MainActivity.java
@@ -6,6 +6,7 @@ import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.view.Window;
+import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
@@ -20,38 +21,58 @@ public class MainActivity extends AppCompatActivity {
         requestWindowFeature(Window.FEATURE_NO_TITLE);
         setContentView(R.layout.login);
 
-      /*  *//*------------------------Tool bar-----------------------*//*
-        Toolbar toolbar = (Toolbar)findViewById(R.id.my_toolbar);
+        Button button = (Button)findViewById(R.id.LoginButton);
+        button.setOnLongClickListener(new View.OnLongClickListener() {
+            @Override
+            public boolean onLongClick(View v) {
+                Intent intent =  new Intent(MainActivity.this,NavigationActivity.class);
 
-        toolbar.setTitleTextColor(Color.parseColor("#ffff33"));
-        toolbar.setSubtitle("Whefe");
-        toolbar.setNavigationIcon(R.mipmap.ic_launcher);
-        setSupportActionBar(toolbar);
-        *//*-------------------------------------------------------*/
+                idEdit = (EditText) findViewById(R.id.LIDEdit);
+                passEdit = (EditText) findViewById(R.id.LPassEdit);
+
+                finish();
+                intent.putExtra("ID_TEXT",idEdit.getText().toString());
+                intent.putExtra("Pass_TEXT",passEdit.getText().toString());
+
+                startActivity(intent);
+
+                return false;
+            }
+        });
+
     }
 
     public void onLoginButtonClicked(View v){
+        boolean success = false;
+
+        /*-------------------------로그인 체크--------------------------*/
 
 
-       Intent intent =  new Intent(MainActivity.this,MapActivity.class);
 
 
-        idEdit = (EditText) findViewById(R.id.LIDEdit);
-        passEdit = (EditText) findViewById(R.id.LPassEdit);
 
 
-        finish();
-        intent.putExtra("ID_TEXT",idEdit.getText().toString());
-        intent.putExtra("Pass_TEXT",passEdit.getText().toString());
+        /*------------------------------------------------------------*/
+        if(success == false){   //로그인 실패
+            Toast.makeText(getApplicationContext(), "로그인 실패", Toast.LENGTH_SHORT).show();
 
-        startActivity(intent);
 
+
+        }else if(success == true){
+            Intent intent =  new Intent(MainActivity.this,NavigationActivity.class);
+
+            idEdit = (EditText) findViewById(R.id.LIDEdit);
+            passEdit = (EditText) findViewById(R.id.LPassEdit);
+
+            finish();
+            intent.putExtra("ID_TEXT",idEdit.getText().toString());
+            intent.putExtra("Pass_TEXT",passEdit.getText().toString());
+
+            startActivity(intent);
+
+        }
     }
 
-    public void onTestButtonClicked(View v){
-        Intent intent =  new Intent(MainActivity.this,OrderActivity.class);
-        startActivity(intent);
-    }
 
     public void onSignupClicked(View v){
         Toast.makeText(getApplicationContext(),"Signup 눌림",Toast.LENGTH_SHORT).show();
