+++ android/MyApplication/app/src/main/res/layout/app_bar_navigation.xml
@@ -1,17 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.chun.whefe.NavActivity">
+    tools:context="com.example.chun.whefe.NavigationActivity"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior">
 
-    <include
-        android:id="@+id/nav_include_map"
-        layout="@layout/map"
-        android:layout_weight="1" />
 
     <android.support.design.widget.AppBarLayout
+        android:id="@+id/nav_appbar"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:theme="@style/AppTheme.AppBarOverlay">
@@ -25,4 +24,16 @@
 
     </android.support.design.widget.AppBarLayout>
 
-</android.support.design.widget.CoordinatorLayout>
+    <fragment
+        android:id="@+id/nav_fragment"
+        android:name="com.example.chun.whefe.fragments.MapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true"
+        android:layout_below="@+id/nav_appbar"
+        android:layout_weight="1" />
+
+    <!--<include layout="@layout/content_navigation" />-->
+
+</RelativeLayout>
