+++ android/MyApplication/app/src/main/res/layout/content_navigation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.example.chun.whefe.NavActivity"
-    tools:showIn="@layout/app_bar_nav">
+    tools:context="com.example.chun.whefe.NavigationActivity"
+    tools:showIn="@layout/app_bar_navigation">
 
 </android.support.constraint.ConstraintLayout>
