+++ android/MyApplication/app/src/main/res/layout/info.xml
@@ -4,52 +4,53 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:layout_weight="1"
+    android:background="@android:color/background_light"
+    app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context="com.example.chun.whefe.InfoActivity">
 
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/my_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentTop="true"
-        android:background="?attr/colorPrimary"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
-
     <RelativeLayout
-        android:id="@+id/relativeLayout"
-        android:layout_width="368dp"
-        android:layout_height="495dp"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintVertical_bias="0.857"
-        android:layout_alignParentBottom="true"
-        android:layout_centerHorizontal="true"
-        android:layout_below="@+id/my_toolbar">
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_below="@+id/my_toolbar"
+        android:layout_alignParentLeft="true"
+        android:layout_alignParentStart="true">
 
         <ScrollView
             android:layout_width="match_parent"
-            android:layout_height="330dp"
+            android:layout_height="match_parent"
             android:layout_above="@+id/OrderButton"
             android:layout_alignParentLeft="true"
             android:layout_alignParentStart="true"
-            android:layout_marginRight="15dp">
+            android:layout_below="@+id/horizontalScrollView">
+
 
             <LinearLayout
                 android:layout_width="match_parent"
                 android:layout_height="300dp"
                 android:orientation="vertical">
 
+                <HorizontalScrollView
+                    android:id="@+id/horizontalScrollView"
+                    android:layout_width="match_parent"
+                    android:layout_height="150dp"
+                    android:layout_marginBottom="10dp">
+
+                    <LinearLayout
+                        android:id="@+id/info_imageList"
+                        android:layout_width="wrap_content"
+                        android:layout_height="match_parent"
+                        android:orientation="horizontal">
+
+                    </LinearLayout>
+                </HorizontalScrollView>
+
                 <TextView
                     android:id="@+id/infoView"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:layout_marginEnd="3dp"
+                    android:layout_marginStart="3dp"
                     android:background="@android:color/holo_orange_light"
                     android:gravity="center"
                     android:text="Null"
@@ -65,6 +66,8 @@
                     android:id="@+id/IntroducevIew"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
+                    android:layout_marginEnd="3dp"
+                    android:layout_marginStart="3dp"
                     android:background="@android:color/holo_green_light"
                     android:paddingBottom="3dp"
                     android:paddingLeft="3dp"
@@ -76,27 +79,6 @@
             </LinearLayout>
         </ScrollView>
 
-        <ImageView
-            android:id="@+id/imageView2"
-            android:layout_width="200dp"
-            android:layout_height="100dp"
-            app:srcCompat="@drawable/grazie1"
-            android:layout_alignParentTop="true"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentStart="true"
-            android:layout_marginTop="11dp" />
-
-        <ImageView
-            android:id="@+id/imageView4"
-            android:layout_width="wrap_content"
-            android:layout_height="100dp"
-            app:srcCompat="@drawable/grazie2"
-            android:layout_alignTop="@+id/imageView2"
-            android:layout_alignParentRight="true"
-            android:layout_alignParentEnd="true"
-            android:layout_toRightOf="@+id/imageView2"
-            android:layout_toEndOf="@+id/imageView2" />
-
         <Button
             android:id="@+id/OrderButton"
             android:layout_width="wrap_content"
@@ -106,20 +88,34 @@
             android:layout_alignParentStart="true"
             android:layout_marginLeft="26dp"
             android:layout_marginStart="26dp"
-            android:onClick="onOrderButtonClicked"
             android:text="메뉴 확인 및 주문" />
 
         <Button
             android:id="@+id/CouponButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:layout_marginLeft="71dp"
+            android:layout_marginStart="71dp"
+            android:text="쿠폰 받기"
             android:layout_alignParentBottom="true"
-            android:layout_marginLeft="22dp"
-            android:layout_marginStart="22dp"
-            android:layout_toEndOf="@+id/imageView2"
-            android:layout_toRightOf="@+id/imageView2"
-            android:onClick="onCouponButtonClicked"
-            android:text="쿠폰 받기" />
-
+            android:layout_toRightOf="@+id/OrderButton"
+            android:layout_toEndOf="@+id/OrderButton" />
     </RelativeLayout>
+
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/my_toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentTop="true"
+        android:background="?attr/colorPrimary"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
+
+
+
+
 </RelativeLayout>
