+++ android/MyApplication/app/src/main/res/layout/order.xml
@@ -4,8 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@android:color/background_light"
     tools:context="com.example.chun.whefe.OrderActivity">
-
     <android.support.v7.widget.Toolbar
         android:id="@+id/my_toolbar"
         android:layout_width="match_parent"
@@ -18,14 +18,13 @@
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent"
         app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
-
     <HorizontalScrollView
         android:id="@+id/hori_scrollview"
         android:layout_width="match_parent"
         android:layout_height="70dp"
-        android:layout_below="@+id/my_toolbar"
         android:layout_alignParentLeft="true"
-        android:layout_alignParentStart="true">
+        android:layout_alignParentStart="true"
+        android:layout_below="@+id/my_toolbar">
 
 
         <LinearLayout
@@ -39,9 +38,9 @@
         android:id="@+id/menuListView"
         android:layout_width="match_parent"
         android:layout_height="380dp"
-        android:orientation="vertical"
         android:layout_above="@+id/shoppingListButton"
-        android:layout_below="@+id/hori_scrollview" />
+        android:layout_below="@+id/hori_scrollview"
+        android:orientation="vertical" />
 
 
     <ImageButton
