+++ android/MyApplication/app/src/main/res/layout/payment.xml
@@ -4,6 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:background="@android:color/background_light"
     tools:context="com.example.chun.whefe.PaymentActivity">
 
     <android.support.v7.widget.Toolbar
@@ -22,16 +23,16 @@
     <RelativeLayout
         android:layout_width="368dp"
         android:layout_height="495dp"
-        android:layout_marginLeft="8dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.857"
-        android:layout_marginRight="8dp"
-        app:layout_constraintRight_toRightOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
         android:layout_alignParentBottom="true"
-        android:layout_below="@+id/my_toolbar">
+        android:layout_below="@+id/my_toolbar"
+        android:layout_marginLeft="8dp"
+        android:layout_marginRight="8dp"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintHorizontal_bias="0.0"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.857">
 
         <Button
             android:id="@+id/paymentButton"
@@ -49,21 +50,22 @@
             android:id="@+id/paymentLinearLayout"
             android:layout_width="match_parent"
             android:layout_height="200dp"
+            android:layout_above="@+id/paymentButton"
             android:layout_below="@+id/payment_explainVIew"
             android:layout_marginLeft="10dp"
             android:layout_marginRight="10dp"
             android:layout_marginTop="10dp"
             android:background="@drawable/layout_border"
             android:orientation="vertical"
-            android:weightSum="1"
-            android:layout_above="@+id/paymentButton">
+            android:weightSum="1">
 
-            <ListView
+            <ExpandableListView
                 android:id="@+id/payment_orderListView"
                 android:layout_width="match_parent"
                 android:layout_height="122dp"
+                android:layout_weight="0.95"
                 android:background="@drawable/layout_border"
-                android:layout_weight="0.95" />
+                android:keepScreenOn="true" />
 
             <TextView
                 android:id="@+id/pay_priceView"
@@ -106,13 +108,13 @@
             android:id="@+id/paymentCancelButton"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginEnd="45dp"
-            android:layout_marginRight="45dp"
-            android:text="취소"
             android:layout_alignBaseline="@+id/paymentButton"
             android:layout_alignBottom="@+id/paymentButton"
+            android:layout_alignEnd="@+id/paymentLinearLayout"
             android:layout_alignRight="@+id/paymentLinearLayout"
-            android:layout_alignEnd="@+id/paymentLinearLayout" />
+            android:layout_marginEnd="45dp"
+            android:layout_marginRight="45dp"
+            android:text="취소" />
 
     </RelativeLayout>
 </RelativeLayout>
