+++ android/MyApplication/app/src/main/res/layout/sh_list_child.xml
@@ -40,7 +40,7 @@
                 android:textSize="14sp" />
 
             <Button
-                android:id="@+id/button3"
+                android:id="@+id/sh_deleteButton"
                 android:layout_width="80dp"
                 android:layout_height="40dp"
                 android:text="삭제" />
