+++ android/MyApplication/app/src/main/res/values/strings.xml
@@ -11,6 +11,8 @@
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 
+
     <string name="str_done">완료</string>
+    <string name="title_activity_navigation">NavigationActivity</string>
 
 </resources>
