+++ android/MyApplication/app/src/main/res/values/styles.xml
@@ -29,15 +29,17 @@
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
     </style>
+
     <style name="AppTheme.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
+
     <style name="Theme.NoShadow" parent="android:style/Theme.Light.NoTitleBar">
         <item name="android:windowContentOverlay">@android:color/transparent</item>
     </style>
 
-    <style name="App.Theme" parent="android:style/Theme.Light.NoTitleBar"/>
+    <style name="App.Theme" parent="android:style/Theme.Light.NoTitleBar" />
 
     <style name="AppTheme.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
